Create separate module for mathematic functions
