+++ scripts/collision/index.js
@@ -1,3 +1,5 @@
+import { clamp } from "../math";
+
 const ObjectType = {
   // NB! Mule is an empty object that allows to create any type of temporary
   // object on a map with either testing purposes or to create one-time objects
@@ -72,9 +74,6 @@ const objectsIntersect = (obj1, obj2) => {
 };
 
 
-const clamp = (val, min, max) => Math.min(max, Math.max(val, min));
-
-
 const updateObject = (obj1) => {
   obj1.y = clamp(obj1.y - obj1.ySpeed, 0, Infinity);
   return obj1;
@@ -82,7 +81,6 @@ const updateObject = (obj1) => {
 
 
 export {
-  clamp,
   createGameObject,
   Direction,
   objectsIntersect,
