+++ scripts/collision/index.test.js
@@ -1,5 +1,4 @@
 import { 
-  clamp,
   createGameObject,
   Direction,
   objectsIntersect,
@@ -7,21 +6,6 @@ import {
   updateObject,
 } from './index';
 
-describe("Utility functions", () => {
-  describe("Clamp function", () => {
-    test("Returns value as is if it's whitin boundaries", () => {
-      expect(clamp(5, 0, 10)).toEqual(5);
-    });
-
-    test("Returns max value if passed value is bigger than max", () => {
-      expect(clamp(15, 0, 10)).toEqual(10);
-    });
-
-    test("Returns min value if passed value is lower than min", () => {
-      expect(clamp(-5, 0, 10)).toEqual(0);
-    });
-  });
-});
 
 describe("Objects intersection", () => {
   test("Objects intersection rectangle is correctly detected", () => {
