+++ scripts/math.js
@@ -1,0 +1,5 @@
+const clamp = (val, min, max) => Math.min(max, Math.max(val, min));
+
+export {
+  clamp,
+};
